% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{anova_data}
\alias{anova_data}
\title{Takes IPD data and returns aggregated time interval data}
\usage{
anova_data(timepoints, timepoints2, ref.study = 1, df)
}
\arguments{
\item{timepoints}{?}

\item{timepoints2}{?}

\item{ref.study}{?}

\item{df}{A data frame of IPD KM data (what variables are required?)}
}
\description{
Takes a generated dataset and formats the data ready to apply the anova parameterisation
}
