% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{auc}
\alias{auc}
\title{Calculate Area Under the Curve}
\usage{
auc(surv, tint = c(1, max(surv[[1]]$time)), quantity = "S")
}
\arguments{
\item{surv}{An object of class `"surv.predicts"` generated by `survcalc()`.}

\item{tint}{A numeric vector of length 2 that indicates the interval over which to
calculate AUC.}

\item{quantity}{specifies the quantity to be plotted. Can be either `"S"` for survival (note that S is upper case),
`"mort"` for mortality, `"haz"` for hazards, `"cumhaz"` for cumulative hazards, `"dH"` for
interval contributions to cumulative hazard.}
}
\value{
Returns a data frame of AUC values for each treatment
}
\description{
Calculate Area Under the Curve
}
\details{
Function uses the `pracma` package for integration, so this must be installed.
}
