% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{calc.risk.tab}
\alias{calc.risk.tab}
\title{Calculates numbers at risk using risk table approach}
\usage{
calc.risk.tab(time, survival, nrisk, tint)
}
\arguments{
\item{time}{X axis coordinates taken from the digitised Kaplan-Meier curve}

\item{survival}{Y axis coordinates taken from the digitised Kaplan-Meier curve; same length as `x`}

\item{nrisk}{A number indicating the number of participants at risk (e.g. N randomised) at time = 0}

\item{tint}{A numeric vector indicating the times at which to estimate number at risk. The
first number must be zero.}
}
\description{
Calculates the number at risk at specific intervals. Assumes that the censoring is random
across the whole range of time.
}
