% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{check.digitised}
\alias{check.digitised}
\title{Checks digitised KM data}
\usage{
check.digitised(digitised)
}
\arguments{
\item{digitised}{A data frame containing variables `time` and `survival`, ordered by
`time`}
}
\description{
Checks a data frame of digitised data to ensure survival and times are
correctly specified (i.e. survival at each time-point is <= survival at previous
time-point and time >= previous time)
}
