% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{clean.digitised}
\alias{clean.digitised}
\title{Cleans digitised KM data}
\usage{
clean.digitised(km)
}
\arguments{
\item{km}{a data frame with two columns that correspond to locations
on the x- (1st column) and y- (2nd column) axes of a digitised Kaplan Meier
curve}
}
\description{
Cleans digitised KM data
}
