% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{combine.ipd.guyot}
\alias{combine.ipd.guyot}
\title{Combines digitised IPD}
\usage{
combine.ipd.guyot(x, y)
}
\arguments{
\item{x}{An object of class `"ipd.guyot"` generated from digitised IPD}

\item{y}{An object of class `"ipd.guyot"` generated from digitised IPD}
}
\value{
Returns a list containing a data frame of IPD that combines data from `x$patient` and
`y$patient`
}
\description{
Combines digitised IPD
}
