% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{cumulative_hazard_function}
\alias{cumulative_hazard_function}
\title{Calculate cumulative hazard function}
\usage{
cumulative_hazard_function(t, params, exponents)
}
\description{
Calculate cumulative hazard function
}
