% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{devplot}
\alias{devplot}
\title{Plot posterior means of residual deviance contributions}
\usage{
devplot(jagsmod1, jagsmod2 = NULL, vline = "study")
}
\arguments{
\item{jagsmod1}{Object of class `rjags` that includes monitored `dev` nodes}

\item{jagsmod2}{Object of class `rjags` that includes monitored `dev` nodes (default is `NULL` but if model
is included then dev-dev plot is created)}

\item{vline}{Can take either `"arm"` or `"study"` to indicate whether vertical dashed lines should indicate change in
arm or study respectively. Can be set as `NULL` to not plot any (will be set to `NULL` for dev-dev plots)}
}
\description{
Plots posterior means of residual deviance contributions or dev-dev plots (if two models are specified)
}
\details{
Note that for dev-dev plots to be created the model must have the same number of deviance contributions. I.e. the
likelihood and data used must be the same. Cannot be used to compare different datasets or likelihoods.
}
