% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{fp_data}
\alias{fp_data}
\title{Prepare JAGS data for fractional polynomial models}
\usage{
fp_data(anova, trtnames)
}
\arguments{
\item{anova}{A list generated by `anova_data()` from digitized KM plots containing two named elements.
`"aggregate"` contains aggregated survival data for use in fractional polynomial models and must have the
following columns:

  * `spgrp` index for the time interval
  * `trialid` index for trial
  * `treatment` treatment names (as character), corresponding to values in `trtnames`
  * `nevents` the number of events in a time interval
  * `natrisk` the number at risk in a time interval
  * `time` time at the end of the time interval (I think)
  * `length` the length (duration) of the time interval}

\item{trtnames}{A character vector of treatment names ordered how the treatment codes should be ordered
within the network}
}
\description{
Prepare JAGS data for fractional polynomial models
}
