% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{fp_geninits}
\alias{fp_geninits}
\title{Create initial values for model}
\usage{
fp_geninits(ns = 4, polyorder = 1, seed = 890421)
}
\arguments{
\item{ns}{An integer for the number of studies in the analysis}

\item{polyorder}{Takes 1 for 1st order and 2 for 2nd order fractional polynomials}

\item{seed}{Can be set to `NULL` if no seed is desired}
}
\description{
Assumes 3 chains
}
