% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{geom_km}
\alias{geom_km}
\title{Adds a geom-ish object for Kaplan-Meier data}
\usage{
geom_km(df, curve_args = list(), ...)
}
\arguments{
\item{df}{A data frame of IPD KM data containing the variables:

* `time` Numeric time at which event/censoring occured
* `event` Binary indicating whether an event (`1`) or censoring (`0`) occured
* `treatment` Factor variable for treatment name
* `study` Factor variable for study name}

\item{curve_args}{Optional list of arguments to customise the curves plotted
with [ggplot2::geom_step()]}

\item{...}{Additional arguments passed to [survival::survfit()]}
}
\description{
Function has been frankensteined from multinma (https://github.com/dmphillippo/multinma/tree/develop)
}
