% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{get_fp}
\alias{get_fp}
\title{Get fractional polynomial parameters}
\usage{
get_fp(x, params, exponents, sums = NULL)
}
\arguments{
\item{x}{A vector with the dependent variable (time).}

\item{params}{A matrix with 2-3 columns (depending on the fractional polynomial order)
giving the intercept and the "slope". If more than one rows given, the fractional polynomial
is evaluated for each row.}

\item{exponent}{A vector of integers giving the exponent(s) (p1 or c(p1,p2)) of the polynomial.}
}
\value{
A matrix with length(x) rows giving the FP values or FP summaries.
}
\description{
Get fractional polynomial parameters
}
\details{
Function copied and edited from gemtcPlus (https://github.com/Roche/gemtcPlus/tree/main)
}
