% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{loglog_plot}
\alias{loglog_plot}
\title{Create log-log plots}
\usage{
loglog_plot(df)
}
\arguments{
\item{df}{A data frame of IPD KM data containing the variables:

* `time` Numeric time at which event/censoring occured
* `event` Binary indicating whether an event (`1`) or censoring (`0`) occured
* `treatment` Factor variable for treatment name
* `study` Factor variable for study name}
}
\description{
Used to assess the validity of the proportional hazards assumption
}
