% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{mcmc_sum}
\alias{mcmc_sum}
\title{Posterior summaries for MCMC}
\usage{
mcmc_sum(mcmc)
}
\arguments{
\item{mcmc}{A numeric vector of MCMC samples (can be a subset of an array/matrix)

Returns a vector}
}
\description{
Posterior summaries for MCMC
}
