% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{print.survquantile}
\alias{print.survquantile}
\title{Prints survival quantiles}
\usage{
\method{print}{survquantile}(x, ...)
}
\arguments{
\item{x}{an object of class `"survquantile"`}

\item{...}{Arguments for `quantile()`}
}
\description{
Prints survival quantiles
}
