% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{sequence_fpoly}
\alias{sequence_fpoly}
\title{Runs multiple fractional polynomial models}
\usage{
sequence_fpoly(
  jagsdat,
  powers = c(-3, -2, -1, -0.5, 0, 0.5, 1, 2, 3),
  polyorder = 1,
  jagsfile = NULL,
  savefile = NULL,
  overwrite = TRUE,
  parameters.to.save = c("d", "mu", "dev", "totresdev"),
  inits = fp_geninits(ns = jagsdat$ns, nt = jagsdat$nt, polyorder = polyorder, seed =
    890421),
  ...
)
}
\arguments{
\item{jagsdat}{A list of data to be read by the model (ideally generated by `fp_data()`).
Must have been generated for the same fractional polynomial model order (`polyorder=1` or `polyorder=2`)
as the set of models to run.}

\item{powers}{A sequence of fractional polynomial powers to run models for. Note that for 2nd order models
all combinations of powers will be run (i.e. lots of models - length(powers) permute 2).}

\item{jagsfile}{The location of a JAGS model saved as a `.jags` file on which to run
the sequence of fractional polynomial models. The default `NULL` uses a standard 1st
or 2nd order (depending on the value of `polyorder`) fractional polynomial model.}

\item{savefile}{The location of a `.rds` file to which the generated object (a list of models) should be saved. Default (`NULL`) does not
save a file.}

\item{overwrite}{Whether the file specified in `savefile` should be overwritten (`TRUE`), or simply appended (`FALSE`)}

\item{...}{Arguments to be sent to JAGS (e.g. `n.iter`)}
}
\value{
A list of fractional polynomial models. Each element is a model (or an error message if the model failed to
run), named as follows `"FP_<1st power>_<2nd power>`.
}
\description{
Runs multiple fractional polynomial models with various power combinations
and can save resulting models. Allows user to set up multiple models to run
back to back (so you can go away and have lunch/sleep/whisky).
}
