% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{sub.ipd.guyot}
\alias{sub.ipd.guyot}
\title{Take a subset of digitised IPD}
\usage{
sub.ipd.guyot(x, y)
}
\arguments{
\item{x}{An object of class `"ipd.guyot"` generated from digitised IPD containing
the larger set of IPD}

\item{y}{An object of class `"ipd.guyot"` generated from digitised IPD that should be
removed from `x`}
}
\description{
Take a subset of digitised IPD
}
\details{
Note that `x` must include patients in `y`. The function then removes IPD patients specified in `y`
from IPD in `x` using nearest-neighbour.
}
