% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{summary.sequence.fpoly}
\alias{summary.sequence.fpoly}
\title{Get a summary of model fit statistics for a sequence of models}
\usage{
\method{summary}{sequence.fpoly}(object, rhat = 1.05, ...)
}
\arguments{
\item{object}{A list of JAGS models of class `"sequence.fpoly` generated by `sequence_fpoly()`}

\item{rhat}{A cutoff for which to make a basic check for convergence. If there are any `mu` or `d`
parameters that have an Rhat greater than this cutoff (default is 1.05) then it will be announced. Can set to
`NULL` which means a convergence check won't be made (and therefore you should do it yourself properly :-P).}

\item{...}{Arguments for `tibble()`}
}
\value{
Returns a tibble of model fit stats. If values are `NA` for a model then it means the model did not run and
returned an error.
}
\description{
Get a summary of model fit statistics for a sequence of models
}
