% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{survquant_chunk}
\alias{survquant_chunk}
\title{Solve for cumulative hazard function at desired survival quantile}
\usage{
survquant_chunk(params, exponents, quantile = 0.5, interval = c(0.1, 1000))
}
\description{
Solve for cumulative hazard function at desired survival quantile
}
