% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{weight_shared_d}
\alias{weight_shared_d}
\title{Weight treatment effects in shared parameter model by proportion of population}
\usage{
weight_shared_d(jagsmod, d.weight)
}
\arguments{
\item{d.weight}{A numeric weighting (between 0 and 1) to apply to treatment
effects for a shared parameter model (e.g. separate treatment effects for
histology). Default (`NULL`) is not to apply any weighting.}
}
\description{
Weight treatment effects in shared parameter model by proportion of population
}
