% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{cumulative_hazard_function}
\alias{cumulative_hazard_function}
\title{Calculate cumulative hazard function}
\usage{
cumulative_hazard_function(
  t,
  lower = 0.1,
  params,
  exponents,
  method = "pracma"
)
}
\arguments{
\item{method}{Can take either `"stats"` or `"pracma"`}
}
\description{
Calculate cumulative hazard function
}
