% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{fp_data}
\alias{fp_data}
\title{Prepare JAGS data for fractional polynomial models}
\usage{
fp_data(data, trtnames, polyorder = 1)
}
\arguments{
\item{data}{A dataframe of aggregate survival data (generated from digitized KM plots).
Must have the following columns:
* `spgrp` index for the time interval
* `trialid` index for trial
* `treatment` treatment names (as character), corresponding to values in `trtnames`
* `nevents` the number of events in a time interval
* `natrisk` the number at risk in a time interval
* `time` time at the end of the time interval (I think)
* `length` the length (duration) of the time interval}

\item{trtnames}{A character vector of treatment names ordered how the treatment codes should be ordered
within the network}

\item{polyorder}{Numeric value indicating the order of the fractional polynomial. Can be either `1` or `2`}
}
\description{
Prepare JAGS data for fractional polynomial models
}
