% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{output_coda_fp}
\alias{output_coda_fp}
\title{Output results from fractional polynomial NMA for use in health economic model}
\usage{
output_coda_fp(
  jagsmod,
  format = "mvn",
  refstudy = "KEYNOTE189",
  treatments = attr(jagsmod, "trtnames")
)
}
\arguments{
\item{jagsmod}{an object of class `"rjags"` that contains the results of a
fractional polynomial model.}

\item{format}{Can take `"mvn"` to output parameters for multivariate normal
distribution, or `"coda"` for a full set of MCMC samples (given as a matrix)}

\item{params}{A vector of parameter names corresponding to those in the model
to extract values for.}
}
\description{
Output results from fractional polynomial NMA for use in health economic model
}
