% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{studykm_survplot}
\alias{studykm_survplot}
\title{Plots survival predictions for each study as baseline for comparison with Kaplan-Meier}
\usage{
studykm_survplot(jagsmod, plotinterval = TRUE)
}
\arguments{
\item{jagsmod}{an object of class `"rjags"` that contains the results of a
fractional polynomial model.}

\item{plotinterval}{indicates whether to plot 95 percent CrI}
}
\description{
Can be used to assess model fit
}
