% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{sub.ipd.guyot}
\alias{sub.ipd.guyot}
\title{Take a subset of digitised IPD}
\usage{
sub.ipd.guyot(x, y, class = "ipd.guyot")
}
\arguments{
\item{x}{An object of class `"ipd.guyot"` or `"data.frame"` (as defined in `class`)
generated from digitised IPD containing
the larger set of IPD}

\item{y}{An object of class `"ipd.guyot"` or `"data.frame"` (as defined in `class`)
generated from digitised IPD that should be removed from `x`}

\item{class}{Can take either `"ipd.guyot"` or `"data.frame"` depending on the
class of `x` and `y`. In future this should just take `"data.frame"`}
}
\description{
Take a subset of digitised IPD
}
\details{
Note that `x` must include patients in `y`. The function then removes IPD patients specified in `y`
from IPD in `x` using nearest-neighbour.
}
