% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{survcalc}
\alias{survcalc}
\alias{plot.surv.predicts}
\title{Predict survival quantities}
\usage{
survcalc(
  jagsmod,
  times = seq(1, 60, length.out = 100),
  refstudy = 1,
  n.mcmc = 2000
)

\method{plot}{surv.predicts}(
  surv,
  quantity = "S",
  treats = surv$treats,
  plotinterval = TRUE,
  kmdat = list()
)
}
\arguments{
\item{jagsmod}{an object of class `"rjags"` that contains the results of a
fractional polynomial model.}

\item{times}{a vector of times at which to estimate time-varying HRs}

\item{refstudy}{index for the study to use as reference when making predictions.}

\item{n.mcmc}{integer to indicate the number of MCMC samples to draw (without replacement). Note that sampling
is random (though is the same across parameters to preserve within-sample correlation). Speeds up computation,
though will result in higher MCMC error.}

\item{surv}{An object of class `"surv.predicts"` generated by `survcalc()`}

\item{quantity}{specifies the quantity to be plotted. Can be either `"S"` for survival (note that S is upper case),
`"mort"` for mortality, `"haz"` for hazards, `"cumhaz"` for cumulative hazards.}

\item{treats}{treatments to plot (as integers). Default plots them all}

\item{plotinterval}{indicates whether to plot 95 percent CrI}

\item{kmdat}{is a list containing data frames of IPD Kaplan-Meier data
to be added to a survival plot. Each data frame should be of a single study arm and should have 1 row
per patient and include the variables `time` (time of event) and `event` (1 for event, 0 for censoring).
List elements should be named by treatment (index) to which they correspond (e.g. list("3"=ipd1, "4"=ipd2)).
Can only be used if `quantity="S"`.}
}
\description{
Outputs survival probabilities (`"S"`), mortality probabilities (`"mort"`),
cumulative hazards (`"cumhaz"`) and hazards (`"haz"`).
}
\section{Functions}{
\itemize{
\item \code{plot(surv.predicts)}: Plot survival predictions

}}
