% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracpoly_functions.R
\name{survcalc}
\alias{survcalc}
\alias{plot.surv.predicts}
\title{Predict survival quantities}
\usage{
survcalc(
  jagsmod,
  refstudy,
  times = seq(1, 60, length.out = 100),
  n.mcmc = 2000
)

\method{plot}{surv.predicts}(
  surv,
  quantity = "S",
  treats = attr(surv, "trtnames"),
  plotinterval = TRUE,
  overlay.km = FALSE
)
}
\arguments{
\item{jagsmod}{an object of class `"rjags"` that contains the results of a
fractional polynomial model.}

\item{refstudy}{Name of the study to use as reference when making predictions. Must match a name in `attr(jagsmod, "studynames")`}

\item{times}{a vector of times at which to estimate time-varying HRs}

\item{n.mcmc}{integer to indicate the number of MCMC samples to draw (without replacement). Note that sampling
is random (though is the same across parameters to preserve within-sample correlation). Speeds up computation,
though will result in higher MCMC error.}

\item{surv}{An object of class `"surv.predicts"` generated by `survcalc()`}

\item{quantity}{specifies the quantity to be plotted. Can be either `"S"` for survival (note that S is upper case),
`"mort"` for mortality, `"haz"` for hazards, `"cumhaz"` for cumulative hazards.}

\item{treats}{treatments to plot (as integers). Default plots them all}

\item{plotinterval}{indicates whether to plot 95 percent CrI}

\item{overlay.km}{Indicates whether Kaplan Meier data for the reference study should be plotted}
}
\description{
Outputs survival probabilities (`"S"`), mortality probabilities (`"mort"`),
cumulative hazards (`"cumhaz"`) and hazards (`"haz"`).
}
\section{Functions}{
\itemize{
\item \code{plot(surv.predicts)}: Plot survival predictions

}}
