% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{survquantile}
\alias{survquantile}
\title{Estimate survival quantiles (e.g. median survival)}
\usage{
survquantile(jagsmod, refstudy, quantile = c(0.5), n.mcmc = NULL)
}
\arguments{
\item{jagsmod}{an object of class `"rjags"` that contains the results of a
fractional polynomial model.}

\item{refstudy}{Name of the study to use as reference when making predictions. Must match a name in `attr(jagsmod, "studynames")`}

\item{n.mcmc}{integer to indicate the number of MCMC samples to draw (without replacement). Note that sampling
is random (though is the same across parameters to preserve within-sample correlation). Speeds up computation,
though will result in higher MCMC error.}

\item{quantiles}{}
}
\description{
Estimate survival quantiles (e.g. median survival)
}
